+++ src/crypto/aws_lc/rsa.rs
@@ -13,7 +13,7 @@ use crate::{Algorithm, DecodingKey, EncodingKey};
 /// Try to sign the `message` using an `RSA` `algorithm`.
 fn try_sign_rsa(
     algorithm: &'static dyn crypto_sig::RsaEncoding,
-    encoding_key: &EncodingKey,
+    encoding_key: &EncodingKey<'_>,
     msg: &[u8],
 ) -> std::result::Result<Vec<u8>, signature::Error> {
     let key_pair = crypto_sig::RsaKeyPair::from_der(encoding_key.inner())
@@ -33,7 +33,7 @@ fn try_sign_rsa(
 /// - If `decoding_key` is not from the RSA family.
 fn verify_rsa(
     algorithm: &'static crypto_sig::RsaParameters,
-    decoding_key: &DecodingKey,
+    decoding_key: &DecodingKey<'_>,
     msg: &[u8],
     signature: &[u8],
 ) -> std::result::Result<(), signature::Error> {
@@ -56,7 +56,7 @@ macro_rules! define_rsa_signer {
         pub struct $name(EncodingKey);
 
         impl $name {
-            pub(crate) fn new(encoding_key: &EncodingKey) -> Result<Self> {
+            pub(crate) fn new(encoding_key: &EncodingKey<'_>) -> Result<Self> {
                 if encoding_key.family() != AlgorithmFamily::Rsa {
                     return Err(new_error(ErrorKind::InvalidKeyFormat));
                 }
@@ -84,7 +84,7 @@ macro_rules! define_rsa_verifier {
         pub struct $name(DecodingKey);
 
         impl $name {
-            pub(crate) fn new(decoding_key: &DecodingKey) -> Result<Self> {
+            pub(crate) fn new(decoding_key: &DecodingKey<'_>) -> Result<Self> {
                 if decoding_key.family() != AlgorithmFamily::Rsa {
                     return Err(new_error(ErrorKind::InvalidKeyFormat));
                 }