+++ src/encoding.rs
@@ -1,4 +1,7 @@
-use std::fmt::{Debug, Formatter};
+use std::{
+    borrow::Cow,
+    fmt::{Debug, Formatter},
+};
 
 use base64::{
     Engine,
@@ -17,32 +20,27 @@ use crate::serialization::{b64_encode, b64_encode_part};
 /// A key to encode a JWT with. Can be a secret, a PEM-encoded key or a DER-encoded key.
 /// This key can be re-used so make sure you only initialize it once if you can for better performance.
 #[derive(Clone)]
-pub struct EncodingKey {
+pub struct EncodingKey<'a> {
     family: AlgorithmFamily,
-    content: Vec<u8>,
+    content: Cow<'a, [u8]>,
 }
 
-impl EncodingKey {
+impl EncodingKey<'_> {
     /// The algorithm family this key is for.
     pub fn family(&self) -> AlgorithmFamily {
         self.family
     }
 
-    /// If you're using a HMAC secret that is not base64, use that.
-    pub fn from_secret(secret: &[u8]) -> Self {
-        EncodingKey { family: AlgorithmFamily::Hmac, content: secret.to_vec() }
-    }
-
     /// If you have a base64 HMAC secret, use that.
-    pub fn from_base64_secret(secret: &str) -> Result<Self> {
+    pub fn from_base64_secret(secret: &str) -> Result<EncodingKey<'static>> {
         let out = STANDARD.decode(secret)?;
-        Ok(EncodingKey { family: AlgorithmFamily::Hmac, content: out })
+        Ok(EncodingKey { family: AlgorithmFamily::Hmac, content: Cow::Owned(out) })
     }
 
     /// For loading websafe base64 HMAC secrets, ex: ACME EAB credentials.
-    pub fn from_urlsafe_base64_secret(secret: &str) -> Result<Self> {
+    pub fn from_urlsafe_base64_secret(secret: &str) -> Result<EncodingKey<'static>> {
         let out = URL_SAFE.decode(secret)?;
-        Ok(EncodingKey { family: AlgorithmFamily::Hmac, content: out })
+        Ok(EncodingKey { family: AlgorithmFamily::Hmac, content: Cow::Owned(out) })
     }
 
     /// If you are loading a RSA key from a .pem file.
@@ -55,10 +53,10 @@ impl EncodingKey {
     /// the key should be at least 2047 bits.
     ///
     #[cfg(feature = "use_pem")]
-    pub fn from_rsa_pem(key: &[u8]) -> Result<Self> {
+    pub fn from_rsa_pem(key: &[u8]) -> Result<EncodingKey<'static>> {
         let pem_key = PemEncodedKey::new(key)?;
         let content = pem_key.as_rsa_key()?;
-        Ok(EncodingKey { family: AlgorithmFamily::Rsa, content: content.to_vec() })
+        Ok(EncodingKey { family: AlgorithmFamily::Rsa, content: Cow::Owned(content.to_vec()) })
     }
 
     /// If you are loading a ECDSA key from a .pem file
@@ -76,35 +74,20 @@ impl EncodingKey {
     ///     | openssl pkcs8 -topk8 -nocrypt -out ec-private.pem
     /// ```
     #[cfg(feature = "use_pem")]
-    pub fn from_ec_pem(key: &[u8]) -> Result<Self> {
+    pub fn from_ec_pem(key: &[u8]) -> Result<EncodingKey<'static>> {
         let pem_key = PemEncodedKey::new(key)?;
         let content = pem_key.as_ec_private_key()?;
-        Ok(EncodingKey { family: AlgorithmFamily::Ec, content: content.to_vec() })
+        Ok(EncodingKey { family: AlgorithmFamily::Ec, content: Cow::Owned(content.to_vec()) })
     }
 
     /// If you are loading a EdDSA key from a .pem file
     /// This errors if the key is not a valid private Ed key
     /// Only exists if the feature `use_pem` is enabled.
     #[cfg(feature = "use_pem")]
-    pub fn from_ed_pem(key: &[u8]) -> Result<Self> {
+    pub fn from_ed_pem(key: &[u8]) -> Result<EncodingKey<'static>> {
         let pem_key = PemEncodedKey::new(key)?;
         let content = pem_key.as_ed_private_key()?;
-        Ok(EncodingKey { family: AlgorithmFamily::Ed, content: content.to_vec() })
-    }
-
-    /// If you know what you're doing and have the DER-encoded key, for RSA only
-    pub fn from_rsa_der(der: &[u8]) -> Self {
-        EncodingKey { family: AlgorithmFamily::Rsa, content: der.to_vec() }
-    }
-
-    /// If you know what you're doing and have the DER-encoded key, for ECDSA
-    pub fn from_ec_der(der: &[u8]) -> Self {
-        EncodingKey { family: AlgorithmFamily::Ec, content: der.to_vec() }
-    }
-
-    /// If you know what you're doing and have the DER-encoded key, for EdDSA
-    pub fn from_ed_der(der: &[u8]) -> Self {
-        EncodingKey { family: AlgorithmFamily::Ed, content: der.to_vec() }
+        Ok(EncodingKey { family: AlgorithmFamily::Ed, content: Cow::Owned(content.to_vec()) })
     }
 
     /// Get the value of the key.
@@ -122,7 +105,29 @@ impl EncodingKey {
     }
 }
 
-impl Debug for EncodingKey {
+impl<'a> EncodingKey<'a> {
+    /// If you're using a HMAC secret that is not base64, use that.
+    pub fn from_secret(secret: &'a [u8]) -> Self {
+        EncodingKey { family: AlgorithmFamily::Hmac, content: Cow::Borrowed(secret) }
+    }
+
+    /// If you know what you're doing and have the DER-encoded key, for RSA only
+    pub fn from_rsa_der(der: &'a [u8]) -> Self {
+        EncodingKey { family: AlgorithmFamily::Rsa, content: Cow::Borrowed(der) }
+    }
+
+    /// If you know what you're doing and have the DER-encoded key, for ECDSA
+    pub fn from_ec_der(der: &'a [u8]) -> Self {
+        EncodingKey { family: AlgorithmFamily::Ec, content: Cow::Borrowed(der) }
+    }
+
+    /// If you know what you're doing and have the DER-encoded key, for EdDSA
+    pub fn from_ed_der(der: &'a [u8]) -> Self {
+        EncodingKey { family: AlgorithmFamily::Ed, content: Cow::Borrowed(der) }
+    }
+}
+
+impl Debug for EncodingKey<'_> {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("EncodingKey")
             .field("family", &self.family)
@@ -153,7 +158,7 @@ impl Debug for EncodingKey {
 /// // This will create a JWT using HS256 as algorithm
 /// let token = encode(&Header::default(), &my_claims, &EncodingKey::from_secret("secret".as_ref())).unwrap();
 /// ```
-pub fn encode<T: Serialize>(header: &Header, claims: &T, key: &EncodingKey) -> Result<String> {
+pub fn encode<T: Serialize>(header: &Header, claims: &T, key: &EncodingKey<'_>) -> Result<String> {
     if key.family != header.alg.family() {
         return Err(new_error(ErrorKind::InvalidAlgorithm));
     }